RESTApi/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy 
pymysql
pydantic
cryptography>=3.4.7
azure-data-tables
python-jose
authlib